+++ lib/app_elecciones/presentation/modules/validate_recinto/local_widget/desing_mapa.dart
@@ -21,6 +21,7 @@ import '../../../../../feactures/mapas/widgets/custom_marker.dart';
 import '../../../../../feactures/user/presentation/modules/controllers.dart';
 import '../../../../data/models/models.dart';
 import '../../../widgets/customWidgets.dart';
+import 'registrar_recinto_widget.dart';
 
 class DesingMapaRecinto extends StatefulWidget {
   final LatLng ubicacion;
@@ -34,6 +35,8 @@ class DesingMapaRecinto extends StatefulWidget {
   final GestureTapCallback ontapMyUbicacion;
   final bool cargando;
 
+  final bool showNoEncuentyroMiRecinto;
+
   const DesingMapaRecinto({
     super.key,
     required this.ubicacion,
@@ -45,6 +48,7 @@ class DesingMapaRecinto extends StatefulWidget {
     this.cargando = false,
     required this.listRecintosElectorales,
     this.onRecintoSeleccionado,
+    this.showNoEncuentyroMiRecinto = false,
   });
 
   @override
@@ -68,6 +72,7 @@ class _DesingMapaRecintoState extends State<DesingMapaRecinto> {
   bool tieneRecintoValidado = false;
 
   Rx<GaleryCameraModel?> mGaleryCameraModel = Rx<GaleryCameraModel?>(null);
+
 
   var controllerNombreRecinto = TextEditingController();
   final formKeyRegRecinto = GlobalKey<FormState>();
@@ -170,13 +175,18 @@ class _DesingMapaRecintoState extends State<DesingMapaRecinto> {
                     return ListView.builder(
                       controller:
                           scrollController, // 👈 IMPORTANTE (NO cambiar)
-                      itemCount: listaOrdenada.length + 1,
+                      itemCount: listaOrdenada.length,
                       itemBuilder: (context, index) {
                         // Último item: No encuentro mi recinto
-                        if (index == listaOrdenada.length) {
+                        if (index == 0) {
                           if (tieneRecintoValidado) {
                             return Container();
                           }
+                          if(widget.showNoEncuentyroMiRecinto==false){
+                            return Container();
+                          }
+
+
                           return Container(
                             margin: EdgeInsets.symmetric(
                               horizontal: 8,
@@ -207,7 +217,7 @@ class _DesingMapaRecintoState extends State<DesingMapaRecinto> {
                                   recintoSeleccionado = null;
                                 });
 
-                                desingRegistrarrecinto();
+                                desingRegistrarRecinto();
                               },
                             ),
                           );
@@ -719,101 +729,21 @@ class _DesingMapaRecintoState extends State<DesingMapaRecinto> {
     );
   }
 
-  void desingRegistrarrecinto() {
-    final responsive = ResponsiveUtil();
+  void desingRegistrarRecinto() {
     DialogosDesingWidget.getDialogoX(
       title: "Nuevo Recinto",
-      contenido: Form(
-        key: formKeyRegRecinto,
+      contenido: RegistrarRecintoWidget(
+        formKey: formKeyRegRecinto,
+        controllerNombreRecinto: controllerNombreRecinto,
+        foto: mGaleryCameraModel,
+        onGuardar: () {
 
-        child: Column(
-          children: [
-            getMsjInformacion(),
-            wgFoto(),
+          ///Guardar recinto
 
- Obx(()=>           mGaleryCameraModel.value != null? Column(children: [
-
-
-   ImputTextWidget(
-     maxLength: 200,
-     keyboardType: TextInputType.multiline,
-     controller: controllerNombreRecinto,
-     icono: Icon(Icons.edit, color: AppColors.colorIcons),
-     label: "Nombre Recinto",
-     fonSize: responsive.diagonalP(AppConfig.tamTextoTitulo),
-     validar: (String? value) {
-       if (value != null && value.length < 5) {
-         return "Ingrese el nombre del Recinto";
-       }
-       return null;
-     },
-   ),
-   BtnIconWidget(
-     icon: Icons.save,
-     titulo: "GUARDAR",
-     onPressed: () {
-       bool isValid = formKeyRegRecinto.currentState!.validate();
-
-       if (!isValid) return;
-     },
-   ),
-
- ],):Container())
-
-          ],
-        ),
+        },
       ),
     );
   }
 
-  Widget getMsjInformacion() {
-    return Container(
-      padding: const EdgeInsets.all(5),
-      decoration: BoxDecoration(
-        color: const Color(0xFFF3F8FF),
-        borderRadius: BorderRadius.circular(15),
-        border: Border(left: BorderSide(color: AppColors.colorAzul, width: 5)),
-      ),
-      child: Row(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          CircleAvatar(
-            radius: 18,
-            backgroundColor: AppColors.colorAzul,
-            child: const Icon(
-              Icons.info_outline,
-              color: Colors.white,
-              size: 20,
-            ),
-          ),
-          const SizedBox(width: 12),
-          Expanded(
-            child: RichText(
-              text: TextSpan(
-                style: const TextStyle(
-                  color: Colors.black87,
-                  fontSize: 14,
-                  height: 1.45,
-                ),
-                children: [
-                  TextSpan(
-                    text: "Importante\n",
-                    style: TextStyle(
-                      fontWeight: FontWeight.bold,
-                      color: AppColors.colorAzul,
-                      fontSize: 15,
-                    ),
-                  ),
-                  const TextSpan(
-                    text:
-                        "Tome una fotografía del afiche oficial del recinto electoral donde se visualice claramente el nombre del recinto.",
-                  ),
-                ],
-              ),
-            ),
-          ),
-        ],
-      ),
-    );
-  }
+
 }