+++ lib/app_elecciones/presentation/modules/validate_recinto/validate_recinto_page.dart
@@ -7,6 +7,7 @@ class ValidateRecintoPage extends GetView<ValidateRecintoController> {
   Widget build(BuildContext context) {
     Widget wg= Obx(
       () => DesingMapaRecinto(
+        showNoEncuentyroMiRecinto: controller.selectProcesosOperativo.permitirAgregarRecintos,
 
     listRecintosElectorales: controller.listRecintosElectorales,
         onPressedSave: () {