+++ lib/feactures/user/presentation/widgets/workAreaPageLoginWidget.dart
@@ -106,7 +106,8 @@ class _WorkAreaPageLoginWidgetState extends State<WorkAreaPageLoginWidget> {
                                 ),
                               ),
                             ))),
-
+// Se agrega un espacio inferior para evitar que
+// la barra de navegación del dispositivo cubra el contenido.
                     SizedBox(
                       height: MediaQuery.of(context).padding.bottom,
                     ),